feat : Date format to 'YYYY-MM'
