+++ wise-office-client/src/types/page.ts
@@ -1,0 +1,11 @@
+export interface PageInfo {
+    currentPage: number;
+    pageSize: number;
+    totalCount: number;
+    totalPages: number;
+}
+
+export interface PageParams {
+    currentPage: number;
+    offset: number;
+}
